Move community icon handling to separate component
